requirements/base.txt
attrs<19.0.0
django>=1.11,<1.12
django-cors-headers>=2.2.0,<2.3
django-environ==0.4.5
django-extensions
django-rest-swagger
django-waffle
djangorestframework
djangorestframework-expander
django-filter==2.0.0
git+https://github.com/alanjds/drf-nested-routers.git
edx-auth-backends
elasticsearch>=6.0.0,<7.0.0
elasticsearch-dsl>=6.0.0,<7.0.0
mysqlclient
pytz
sqlparse
